Create README.md
Hello! I'm Michael Mungai, a data enthusiast with a strong background in applied statistics and data analytics. I'm passionate about leveraging artificial intelligence for academia and to solve healthcare challenges. With an MSc in Data Analytics from KCA University and a BSc in Applied Statistics with IT from Masinde Muliro University of Science and Technology, I bring a solid foundation in both theory and practice.